+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.9.0
 -----
 
-- Add support for integrating external Sphinx projects with :confval:`visualised_projects`
+- Add support for integrating external Sphinx projects with :confval:`visualized_projects`
 
 0.8.2
 -----